Figured out AJAx calls, added in method to save and load general configuration settings
Methods currently described on SSID_Page.cs though maybe should move
them elsewhere. They depend on the Util.hs function which I'm still not
sure if this is example plugin specific or a general homeseer function.
Currently saves all configurations everytime the ajax call is made, and
currently makes the call every keystroke (in the case of the time in
miliseconds), may add throttles later if this is a problem.