Attempt to replicate MODBUS addDevice page
Modbus adddevice page seems to be the Homeseer add-device page but is
different. It looks like it overrides the real one because when the
modbus plugin is active it takes over the + button on the home tab.
Seems a bit like Modbus people recreated the Add-Devices pages for their
plugin because it's slightly different. So I've added some placeholder
html to make a version of the add device page for the Modbus aspect of
our SIID plugin.  Need to figure out how to add devices for real and
make the page populate with the devices attributes so I can use it again
for editing the device.

Also moved some functions into more appropriate .cs files so thei're a
bit more logical in placement.