Added modbus device info to added device
Attached the created device to the plugin. Added a switch to the plugin
to check if the device is a modbus device and then to render the modbus
device settings int he plugin's tab. Plugin device information is saved
as a http collection string in the Device Type String field. Modbus
gateways are under Type=Modbus+Gateway

Modbus devices and Bacnet or whatever are to be determined.

Added a ajax callback which takes modbus edits to the plugin tab and
makes the changes in the device type string collection.

Next step, display the created gateways on the plugin modbus devices
page, then do Modbus devices which are not gateways, then start using
these settings and whatnot to do actuall modbus communications.