Made device tab update live when type changes. Added way tor emove deleted devices from gateway's device list
When the plugin's modbus config page populates, it iterates through all
the gateway's list of devices, if the device is null it removes the
entry from the list.

The device page upadates by showing and hiding device configuration rows
as appropriate, in page JS is used to update which rows are to be
displayed.