Added NModbus4 package, Started addding infastructure for writing/reading registers
Added more types. Coils and discrete inputs are 1 bit and registers are
16 bit (single is 2char or a 16-int) and can read multiple to a wider
array of datatypes.

Got a working test going with the sem3 device. Other tools can
read/write to holding register 60021 (40000+20021)  Made me realize that
the existing modbus plugin kind of sucks. Wont be hard to make something
that works better.  Need to figure out what to do with red modbus device
as I can't really connect to it.