Have working export, working on import
Have working import button which uploads a csv file, and sends it in an
ajax request to the backend. Need to add a loading file mask, add a
parsing file mask. need to parse the csv and create the devices. With
those done then Import and Export should be checked off pendning user
feedback on what more they want, some parts of devices are not exported
because I'm not sure how to retrive them, Import is planned on
automatically setting some defaults based on export gaps and on the fact
that we're definitely not going to export and import images and action
groups in this way.

May need to look into exporting and importing action control groups
because that would speed up house deployment.

Yay for Peers work making these import to/from csv things I've done many
times before.