Added controll actions for scratchpad rules for Steward
Also added a switch for setIOmulti, currently only checks if the action
is a scratchpad action or a modbus device action, spins off a thread to
handle the action.

Steward will send API commands through json to scratchpad devices. If
number is negative, triggers an accumulator reset. if it's positive will
set it to enabled, accumulator, type=monthly and day to the value sent.