Upgraded calc parser to obey order of operations, Added scratchpad sub rules
Using now SoftCircuits eval function from
http://www.blackbeltcoder.com/Articles/algorithms/a-c-expression-evaluator

Now users can add a scratchpad sub rule to store things like meter rates
or whatever. The subrule is linked to the rule by the associated devices
so steward can see it. Commands so subrules will set the value of the
subrule to the command value so steward can change it.