3.0.0.10
This hopefully solves the issues that have been persistent with imported
Modbus devices not showing up in the SIID page, and will also allow
import/export of device status controls and custom status icons.

Changes:
-Scratchpad rule rate information and realtime device ID properly
exported and imported.
-SIID device control and status graphics information properly exported
and imported.
-Modbus devices without a gateway now display in the SIID page as
"Orphaned Devices" and the gateway can be set on them appropriately.
This was one of the issues with imports failing.